backend/requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg[binary]>=3.2.0
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# File Upload & Storage
cloudinary==1.41.0
Pillow>=10.0.0

# API Documentation
drf-spectacular==0.27.1

# Filtering
django-filter==23.5

# Geolocation (requires GDAL installed system-wide)
# Will add after GDAL setup

# Utils
python-slugify==8.0.2
pytz==2024.1
requests==2.31.0

# Validation
email-validator==2.1.0
